+++ src/main/java/com/hotel/hotel_backend/dto/response/PaymentSessionResponse.java
@@ -7,7 +7,7 @@ public record PaymentSessionResponse(
         Long bookingId,
         String method,
         String status,
-        Double amount,
+        Long amount,
         String paymentCode,
         String transferContent,
         String bankAccountNo,
